[MAINT] fix #641 - split cli into multiple modules.
